+++ resources/views/dashboard/admin/auth/permission/edit.blade.php
@@ -1,0 +1,8 @@
+<x-layouts.dashboard>
+    <x-slot:title>
+        {{__("string. Permissions")}}
+    </x-slot:title>
+
+    <livewire:admin.permissions.edit :id="$id" />
+
+</x-layouts.dashboard>