+++ resources/views/dashboard/admin/auth/permission/index.blade.php
@@ -1,0 +1,8 @@
+<x-layouts.dashboard>
+    <x-slot:title>
+        {{__("string. Permissions")}}
+    </x-slot:title>
+
+    <livewire:data-tables.permission-data-table />
+
+</x-layouts.dashboard>